+++ src/ast/block/attribute/mod.rs
@@ -1,5 +1,5 @@
-pub use self::name::AttributeName;
-pub use self::value::AttributeValue;
+pub use self::name::{AttributeName, InvalidAttributeName};
+pub use self::value::{AttributeValue, InvalidAttributeValue};
 
 /// Convenience type alias.
 pub type Attribute<'a> = (AttributeName<'a>, Option<AttributeValue<'a>>);