+++ src/ast/mod.rs
@@ -1,8 +1,9 @@
 pub use self::block::BlockName;
-pub use self::block::{Attribute, AttributeName, AttributeValue, Block};
-pub use self::error::IllegalChar as IllegalCharError;
+pub use self::block::{
+    Attribute, AttributeName, AttributeValue, Block, InvalidAttributeName, InvalidAttributeValue,
+    InvalidBlockName,
+};
 pub use self::section::Section;
 
 mod block;
-mod error;
 mod section;